% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{aapl}
\alias{aapl}
\title{Prix de l'action Apple}
\format{
Une base de données avec 313 lignes et 7 variables:
\describe{
  \item{\code{Date}}{\code{Date}}
  \item{\code{Open}}{double, prix à l'ouverture}
  \item{\code{High}}{double, prix maximal quotidien}
  \item{\code{Low}}{double, prix minimal quotidien}
  \item{\code{Close}}{double, prix à la fermeture}
  \item{\code{Volume}}{double,  quantité d'actions transigées représentant volume de transaction}
  \item{\code{Adj_Close}}{double, prix ajusté de fermeture}
}
}
\usage{
aapl
}
\description{
Prix de l'action d'Apple
}
\keyword{datasets}
