% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{abonnement}
\alias{abonnement}
\title{Abonnement de télécommunications}
\format{
Une base de données avec 1000 lignes et 7 variables:
\describe{
  \item{\code{duree}}{nombre de jours d’abonnement}
  \item{\code{tchange}}{nombre de jours d’abonnement au moment du changement de forfait}
  \item{\code{actif}}{entier, variable binaire codée 0 si le client est inactif, 1 s’il est encore abonné}
  \item{\code{nserv1}}{entier, nombre de services lors du contrat initial}
  \item{\code{nserv2}}{entier, nombre de services lors du changement de forfait le cas échéant (valeur manquante sinon)}
  \item{\code{age}}{facteur pour la tranche d’âge, soit (18,35], (35,50] et (50,90] ans}
  \item{\code{sexe}}{variable indicatrice binaire, soit 0 pour femmes et 1 pour hommes}
}
}
\usage{
abonnement
}
\description{
Données fictives de durées d’abonnement chez un fournisseur de services de télécommunications (télévision, internet, téléphonie fixe et cellulaire).
}
\author{
Léo Belzile
}
\keyword{datasets}
