% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{adjustement_factanal}
\alias{adjustement_factanal}
\title{Critères d'information pour analyse factorielle}
\usage{
adjustement_factanal(factors, ...)
}
\arguments{
\item{factors}{le nombre de facteurs à ajuster (vecteur)}

\item{...}{autres arguments passés à factanal}
}
\description{
Cette fonction calcule les estimation
du maximum de vraisemblance pour une
analyse factorielle.
}
