% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{aerien}
\alias{aerien}
\title{Enquête de satisfaction d'une compagnie aérienne}
\format{
Une base de données avec 129880 lignes et 23 variables:
\describe{
  \item{\code{sexe}}{facteur, sexe (\code{homme} ou \code{femme})}
  \item{\code{loyaute_consommateur}}{facteur, est-ce que la personne a un compte fidélité}
  \item{\code{age}}{entier, âge de la personne}
  \item{\code{type_deplacement}}{facteur, motif du déplacement (\code{affaires} ou \code{personnel})}
  \item{\code{classe}}{facteur, classe du vol (\code{affaire}, \code{eco} ou \code{ecoplus})}
  \item{\code{distance_vol}}{double, distance du vol (en km)}
  \item{\code{service_internet_en_vol}}{entier, niveau de satisfaction pour le service internet en vol (0 si pas applicable)}
  \item{\code{temps_arrivee_depart_convenable}}{entier, niveau de satisfaction pour l’heure de départ et d’arrivée}
  \item{\code{facilite_reservation_en_ligne}}{entier, niveau de satisfaction en rapport à la facilité d’utilisation du service de réservation en ligne}
  \item{\code{localisation_porte}}{entier, niveau de satisfaction en lien avec la localisation de la porte d’embarquement}
  \item{\code{nourriture}}{entier, niveau de satisfaction en lien avec la nourriture et les breuvages servis en vol}
  \item{\code{preenregistrement_en_ligne}}{entier, niveau de satisfaction en lien avec le pré-enregistrement en ligne}
  \item{\code{confort_siege}}{entier, niveau de satisfaction lié au confort du siège}
  \item{\code{divertissement_en_vol}}{entier, niveau de satisfaction quant à l’offre de divertissement en vol}
  \item{\code{service_embarquement}}{entier, niveau de satisfaction sur l’embarquement du vol}
  \item{\code{service_espace_jambes}}{entier, niveau de satisfaction en lien avec l’espace pour les jambes}
  \item{\code{service_enregistrement}}{entier, niveau de satisfaction lié au service d’enregistrement à la porte}
  \item{\code{service_vol}}{entier, niveau de satisfaction en lien avec le service durant le vol}
  \item{\code{proprete}}{entier, niveau de satisfaction pour la propreté de l’appareil}
  \item{\code{delai_depart_min}}{entier, retard du vol au départ (en minutes)}
  \item{\code{delai_arrivee_min}}{entier, retard du vol à l'arrivée (en minutes)}
  \item{\code{satisfaction}}{facteur, indice binaire de satisfaction globale}
}
}
\usage{
aerien
}
\description{
Résultats d'une enquête de satisfaction sur des clients de compagnie aérienne. Les variables des colonnes 7 à 20 sont tirées des résultats du questionnaire, les autres encodent des informations démographiques sur les clients et sur le vol.
}
\details{
Les variables mesurent la satisfaction (mesurée par une échelle
de Likert allant de 1 à 5, où un score plus élevé témoigne d’une plus
grande satisfaction. Les valeurs de 0 correspondent à des champs non
applicables (une forme de données manquantes).

Les données sont mises à disposition sous licence CC0 (domaine public) issues de Kaggle.
}
\references{
Kaggle \url{https://www.kaggle.com/sjleshrac/airlines-customer-satisfaction/}
}
\author{
Sayantan Jana
}
\keyword{datasets}
