% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{AIC.regsubsets}
\alias{AIC.regsubsets}
\title{Critère d'information d'Akaike}
\usage{
\method{AIC}{regsubsets}(object, ..., k = 2)
}
\arguments{
\item{object}{un objet de classe \code{regsubsets}}

\item{...}{arguments additionnels, actuellement ignorés}

\item{k}{valeur de pénalité, défaut à 2 par paramètres}
}
\value{
vecteur avec critère d'information pour chaque modèle
}
\description{
Cette fonction retourne le critère d'information, tel que calculé pour le BIC
pour un objet de classe \code{regsubsets}, en supposant que tous les modèles de
1 à \code{nvmax} variables sont retournés.
}
