% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{alpha}
\alias{alpha}
\title{Alpha de Cronbach}
\usage{
alpha(x)
}
\arguments{
\item{x}{matrice de variables numériques}
}
\value{
alpha de Cronbach
}
\description{
Calcul de la mesure de cohérence interne
pour la création d'échelles
}
