% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{alpha_Cronbach}
\alias{alpha_Cronbach}
\title{Alpha de Cronbach}
\usage{
alpha_Cronbach(x)
}
\arguments{
\item{x}{matrice de variables numériques}
}
\value{
alpha de Cronbach
}
\description{
Calcul de la mesure de cohérence interne
pour la création d'échelles
}
