% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{bartlettcorr_factanal}
\alias{bartlettcorr_factanal}
\title{Correction de Bartlett pour analyse factorielle}
\usage{
bartlettcorr_factanal(object)
}
\arguments{
\item{object}{objet de class \code{factanal}}
}
\value{
une liste avec les éléments suivants:
\describe{
\item{\code{statistic}}{ statistique du rapport de vraisemblance fois correction de Bartlett}
\item{\code{pvalue}}{valeur-p pour la statistique modifiée}
}
}
\description{
Calcule une correction de Bartlett pour l'analyse factorielle
}
