% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{BIC.regsubsets}
\alias{BIC.regsubsets}
\title{Critère d'information BIC}
\usage{
\method{BIC}{regsubsets}(object, ...)
}
\arguments{
\item{object}{un objet de classe \code{regsubsets}}

\item{...}{arguments additionnels, actuellement ignorés}
}
\value{
vecteur avec critère d'information pour chaque modèle
}
\description{
Cette fonction retourne le critère d'information BIC
}
