% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{biere}
\alias{biere}
\title{Classement de bières}
\format{
Une base de données avec 35 lignes et 7 variables:
\describe{
  \item{\code{classement}}{entier, évaluation subjective de la qualité de la bière, soit très bon (1), bon (2) ou correct (3).}
  \item{\code{biere}}{facteur, type de bière}
  \item{\code{cout}}{double, coût pour 12 onces}
  \item{\code{calories}}{double, nombre de calories}
  \item{\code{sodium}}{double, sodium en milligrammes}
  \item{\code{alcool}}{double, pourcentage d’alcool}
}
}
\usage{
biere
}
\description{
Classement de bières
}
\author{
Source inconnue, données fournies par Julie Meloche
}
\keyword{datasets}
