% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{bjffacto}
\alias{bjffacto}
\title{Douleur comme liant social}
\format{
Une base de données avec 54 lignes et 21 variables:
\describe{
 \item{\code{condition}}{entier, indicateur binaire du groupe expérimental (0 pour groupe contrôle, 1 pour douleur)}
 \item{\code{interested}}{entier, échelle de Likert pour intéressé(e)}
 \item{\code{distressed}}{entier, échelle de Likert pour angoissé(e)}
 \item{\code{excited}}{entier, échelle de Likert pour excité(e)}
 \item{\code{upset}}{entier, échelle de Likert pour fâché(e)}
 \item{\code{strong}}{entier, échelle de Likert pour fort(e)}
 \item{\code{guilty}}{entier, échelle de Likert pour coupable}
 \item{\code{scared}}{entier, échelle de Likert pour effrayé(e)}
 \item{\code{hostile}}{entier, échelle de Likert pour hostile}
 \item{\code{enthusiastic}}{entier, échelle de Likert pour enthousiaste}
 \item{\code{proud}}{entier, échelle de Likert pour fier/fière}
 \item{\code{irritable}}{entier, échelle de Likert pour irritable}
 \item{\code{alert}}{entier, échelle de Likert pour alerte}
 \item{\code{ashamed}}{entier, échelle de Likert pour honteux/honteuse}
 \item{\code{inspired}}{entier, échelle de Likert pour inspiré(e)}
 \item{\code{nervous}}{entier, échelle de Likert pour nerveux/nerveuse}
 \item{\code{determined}}{entier, échelle de Likert pour déterminé(e)}
 \item{\code{attentive}}{entier, échelle de Likert pour attentif/attentive}
 \item{\code{jittery}}{entier, échelle de Likert pour agité(e)}
 \item{\code{active}}{entier, échelle de Likert pour actif/active}
 \item{\code{afraid}}{entier, échelle de Likert pour craintif/craintive}
}
}
\usage{
bjffacto
}
\description{
Données tirées de l’article Bastian et al. (2014) et rattachées à une
expérience de psychologie visant à corroborer l’hypothèse qu’il y a une coopération
accrue entre individus sujets à une expérience traumatisante. La moitié des
participant(e)s a dû plonger sa main dans un bain d’eau glacé, tandis que
l’autre moitié a dû faire la même chose dans un bain d’eau tiède; les deux groupes
 devaient ensuite faire un jeu visant à identifier leur niveau de coopération.
 Dans la phase initiale, les participant(e)s devaient évaluer leur état d’esprit
 sur une échelle de Likert allant de 1 (peu ou pas du tout) à 5 (extrêmement) en
 indiquant leurs émotions.
}
\references{
Bastian, B., Jetten, J., & Ferris, L. J. (2014). \emph{Pain as social glue : Shared pain increases cooperation}. Psychological Science, \bold{25}(11), 2079–2085.
}
\keyword{datasets}
