% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistique.R
\name{calibration}
\alias{calibration}
\title{Test de calibration de Spiegelhalter}
\usage{
calibration(prob, resp, ...)
}
\arguments{
\item{prob}{vecteur de prédictions (probabilités)}

\item{resp}{vecteur d'observations (variables binaires) ou facteur avec niveaux 0 et 1.}

\item{...}{autres arguments, actuellement ignorés}
}
\value{
une liste avec éléments
\itemize{
\item{stat}{statistique du score}
\item{pval}{valeur-p}
}
}
\description{
Le test de Spiegelhalter est approximativement
standard normal en grand échantillon et
 est basé sur une décomposition d'un
 critère de score.
}
\references{
D.J. Spiegelhalter (1986). \emph{Probabilistic prediction in patient management and clinical trials}, Statistics in Medecine, \bold{5}(5), pp. 421-433, \doi{doi.org/10.1002/sim.4780050506}.
}
\author{
Leo Belzile
}
