% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{chaussures}
\alias{chaussures}
\title{Ventes d'un magasin de chaussures}
\format{
Une base de données avec 6807 lignes et 4 variables:
\describe{
 \item{\code{statut}}{entier, variable catégorielle valant 0 si les chaussures sont vendues, 1 si elles sont toujours en stock, 2 si elles sont déstockées.}
 \item{\code{temps}}{double temps de stockage de l’article (en mois).}
 \item{\code{sexe}}{entier, variable indicatrice binaire, 0 pour modèle pour homme, 1 pour femme.}
 \item{\code{prix}}{double prix de vente réelle de l’article (avec rabais si applicable), arrondi à l’unité près.}
}
}
\usage{
chaussures
}
\description{
Données simulées d'inventaire de chaussures d'un magasin montréalais. Les modèles invendus après 40 mois en magasins sont passés aux pertes et profits. Pour éliminer les invendus lors de l’arrivée de nouveaux modèles, l’entreprise offre une réduction
de 20% après 15 mois.
}
\author{
Léo Belzile
}
\keyword{datasets}
