% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{chomage}
\alias{chomage}
\title{Étude sur le chômage}
\format{

}
\usage{
chomage
}
\description{
Dans le cadre d’un programme gouvernemental,
 on s’intéresse au nombre de semaines de paiements écoulé
 entre une perte d’emploi et un nouvel emploi pour des
 chômeurs et chômeuses âgé(e)s entre 18 et 65 ans,
 dans le but d’élaborer un programme de formation destiné
 au chômage de longue durée. La base de données contient
 un échantillon fictif de 2 500 personnes inscrites
 à l’assurance emploi entre janvier et novembre 2021
}
\author{
Léo Belzile
}
\keyword{datasets}
