% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{coefscore_factanal}
\alias{coefscore_factanal}
\title{Coefficients du score pour analyse factorielle}
\usage{
coefscore_factanal(object, ...)
}
\arguments{
\item{object}{objet de class \code{factanal}}

\item{...}{actuellement ignoré}
}
\description{
Calcul des coefficients de régression
associés aux scores factoriels.
}
