% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{colinearite}
\alias{colinearite}
\title{Données colinéaires}
\format{
Une base de données avec 100 lignes et 6 variables:
\describe{
 \item{\code{y}}{entier, variable réponse binaire}
 \item{\code{x1}}{double, variable explicative}
 \item{\code{x2}}{double, variable explicative}
 \item{\code{x3}}{double, variable explicative}
 \item{\code{x4}}{double, variable explicative}
 \item{\code{x5}}{double, variable explicative}
}
}
\usage{
colinearite
}
\description{
Données simulées pour illustrer la colinéarité dans le cadre d'une régression logistique
}
\keyword{datasets}
