% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{college}
\alias{college}
\title{Admission dans un collège}
\format{
Une base de données avec 777 lignes et 19 variables:
\describe{
 \item{\code{prive}}{variable binaire indiquant si l’institution d’enseignement est privée (\code{1}) ou pas (\code{0})}
 \item{\code{napplications}}{nombre d’applications pour les études}
 \item{\code{nadmission}}{nombre de personnes admises parmi napplications}
 \item{\code{ninscrits}}{nombre d’étudiant(e)s inscrit(e)s à l’université}
 \item{\code{m10p}}{pourcentage des admis provenant d’une école secondaire classé parmi les 10 pourcent meilleures}
 \item{\code{m25p}}{pourcentage des admis provenant d’une école secondaire classé parmi les 25 pourcent meilleures}
 \item{\code{tempsplein1c}}{nombre d’étudiant(e)s à temps plein au premier cycle}
 \item{\code{tempspart1c}}{nombre d’étudiant(e)s à temps partiel au premier cycle}
 \item{\code{fraiscolexternes}}{frais de scolarité pour étudiant(e)s externes (hors état)}
 \item{\code{fraisres}}{frais pour les résidences et l’hébergement}
 \item{\code{fraislivres}}{frais moyens pour les ouvrages obligatoires}
 \item{\code{fraisperso}}{frais pour des dépenses personnelles}
 \item{\code{pourcentdoctorat}}{taux de membres du corps enseignant détenteurs d’un Ph.D.}
 \item{\code{pourcentterminal}}{taux de membres du corps enseignant détenteurs d’un diplôme terminal}
 \item{\code{ratioetudprof}}{rapport du nombre d’étudiant(e)s versus professeur(re)s}
 \item{\code{pourcentdonationdiplome}}{pourcentage des diplômé(e)s qui font des dons à l’institution}
 \item{\code{depenseparetud}}{dépenses liées à l’enseignement, par étudiant(e)}
 \item{\code{tauxdiplom}}{taux de diplomation}
 \item{\code{nom}}{nom de l'établissement}
}
}
\usage{
college
}
\description{
Données d'admission de collèges et d'universités aux États-Unis
}
\keyword{datasets}
