% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{creation_echelles}
\alias{creation_echelles}
\title{Création d'échelles}
\usage{
creation_echelles(chargements, data, seuil = 0.4)
}
\arguments{
\item{chargements}{matrice de chargements, \code{loadings}}

\item{data}{matrice ou base de données}

\item{seuil}{valeur du seuil pour les chargements}
}
\value{
une liste avec les échelles, \code{echelles}, et les valeurs du alpha de Cronbach, \code{alpha}
}
\description{
Calcul de moyennes de colonnes à partir d'une base de données,
en regroupant les variables dont les chargements excèdent le seuil
}
