% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistique.R
\name{.check_resp}
\alias{.check_resp}
\title{Vérifications d'usage pour une variable binaire}
\usage{
.check_resp(resp)
}
\arguments{
\item{resp}{un vecteur numérique, binaire, logique ou un facteur}
}
\value{
un message d'erreur ou vecteur entier avec valeurs 0L ou 1L
}
\description{
Vérifications d'usage pour une variable binaire
}
\keyword{internal}
