% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{eboulis}
\alias{eboulis}
\title{Diagramme d'éboulis}
\usage{
eboulis(object, which = 1:2)
}
\arguments{
\item{object}{objet de classe \code{eigen}, \code{princomp} ou \code{prcomp}}

\item{which}{vecteur d'entiers indiquant quels graphiques imprimer, soit \code{1} pour le diagramme d'éboulis et \code{2} pour la variance cumulative}
}
\value{
une matrice de deux objets \code{ggplot}
}
\description{
Crée un diagramme d'éboulis représentant les valeurs propres d'une matrice de covariance, soit les corrélations ou les variances des variables, en ordre décroissant.
}
