% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{elnino}
\alias{elnino}
\title{El Nino}
\format{
Une base de données avec 10000 lignes et 12 variables:
\describe{
 \item{\code{obs}}{numéro d'observation}
 \item{\code{year}}{année}
 \item{\code{month}}{mois}
 \item{\code{day}}{jour}
 \item{\code{date}}{date}
 \item{\code{latitude}}{latitude}
 \item{\code{longitude}}{longitude}
 \item{\code{zon_winds}}{vents (composante zonale, négatif pour ouest, positive pour est}
 \item{\code{mer_winds}}{vents (composante méridionale, négatif pour sud, positif pour nord)}
 \item{\code{humidity}}{pourcentage d'humidité relative}
 \item{\code{air_temp}}{température ambiante}
 \item{\code{s_s_temp}}{températures de la surface de la mer}
}
}
\source{
Diane Cook, sous-ensemble des données disponibles sur le \href{https://archive.ics.uci.edu/ml/datasets/El+Nino}{UCI Machine Learning repository}
}
\usage{
elnino
}
\description{
Données météorologiques et de surface obtenues à l'aide de bouées positionnées dans le Pacifique équatorial
}
\keyword{datasets}
