% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{eval_EQM_regsubsets}
\alias{eval_EQM_regsubsets}
\title{Calcul de la racine de l'erreur moyenne quadratique}
\usage{
eval_EQM_regsubsets(model, select = c("AIC", "BIC"), formula, data, newdata)
}
\arguments{
\item{model}{objet de classe \code{regsubsets}}

\item{select}{vecteur des modèles sélectionnés}

\item{formula}{formule désignant le modèle maximal ajusté par la procédure de sélection}

\item{data}{base de données d'entraînement servant à l'ajustement des coefficients}

\item{newdata}{base de données de validation pour le calcul de l'erreur moyenne quadratique}
}
\value{
estimation de la racine de l'erreur quadratique moyenne
}
\description{
Pour un objet de class \code{regsubsets},
calculer l'erreur quadratique moyenne de validation
pour une base de données externe. Le modèle calcule
la matrice du modèle pour extraire les colonnes.
}
