% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{factocp}
\alias{factocp}
\title{Analyse factorielle avec composantes principales}
\usage{
factocp(x, covmat = NULL, cor = TRUE, nfact = "kaiser")
}
\arguments{
\item{x}{une matrice ou base de données}

\item{covmat}{matrice de covariance}

\item{cor}{logique; si \code{TRUE}, calculer la décomposition à partir de la matrice de corrélation plutôt que la matrice de covariance}

\item{nfact}{entier pour le nombre de facteurs, ou chaîne de caractère \code{"kaiser"} pour le critère des valeurs propres.}
}
\value{
une liste avec composante \code{loadings} contenant les chargements et si le nombre de facteurs est supérieur à un, la matrice de rotation varimax.
}
\description{
Cette fonction calcule une décomposition spectrale d'une matrice de corrélation ou de covariance des données et extrait les k premières composantes selon \code{nfact}, soit un entier ou le critère de Kaiser.
}
\examples{
factocp(covmat = cov(factor), nfact = 2)
factocp(factor, nfact = "kaiser", cor = FALSE)
}
