% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{factor}
\alias{factor}
\title{Étude dans un magasin}
\format{
Une base de données avec 200 lignes et 12 variables:
\describe{
 \item{\code{x1}}{le magasin offre de bons prix tous les jours?}
 \item{\code{x2}}{le magasin accepte les cartes de crédit majeures (Visa, Mastercard)?}
 \item{\code{x3}}{le magasin offre des produits de qualité?}
 \item{\code{x4}}{les vendeurs connaissent bien les produits?}
 \item{\code{x5}}{il y ait des ventes spéciales régulièrement?}
 \item{\code{x6}}{les marques connues soient disponibles?}
 \item{\code{x7}}{le magasin ait sa propre carte de crédit?}
 \item{\code{x8}}{le service soit rapide?}
 \item{\code{x9}}{il y ait une vaste sélection de produits?}
 \item{\code{x10}}{le magasin accepte le paiement par carte de débit?}
 \item{\code{x11}}{le personnel soit courtois?}
 \item{\code{x12}}{ le magasin ait en stock les produits annoncés?}
}
}
\usage{
factor
}
\description{
Pour les besoins d’une enquête, on a demandé à 200 consommateurs adultes de répondre aux questions suivantes par rapport à un certain type de magasin sur une échelle de Likert de 1 (pas important) à 5 (très important) en jugeant de l'importance de 12 items.
}
\author{
Denis Larocque
}
\keyword{datasets}
