% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroupements.R
\name{homogeneite}
\alias{homogeneite}
\title{Coefficient de détermination (semi-partiel)}
\usage{
homogeneite(
  scd = NULL,
  rhier = NULL,
  data = NULL,
  ngroupes = length(scd),
  which = 1:2
)
}
\arguments{
\item{scd}{somme totale du carré des distances intra-groupes pour chaque nombre de regroupements}

\item{rhier}{objet de classe \code{hclust}; si présent, les valeurs de \code{scd} sont ignorées}

\item{data}{base de données si l'argument \code{rhier} est fourni}

\item{ngroupes}{entier indiquant le nombre de regroupements maximum}

\item{which}{vecteur d'entiers indiquant quels graphiques imprimer, soit \code{1} pour le diagramme du coefficient de déterminationet \code{2} pour le coefficient de détermination semi-partiel}
}
\value{
une matrice de deux objets \code{ggplot}
}
\description{
Crée un diagramme du coefficient de détermination et du coefficient de détermination semi-partielle pour les modèles de regroupements hiérarchiques de classe \code{hclust} avec la méthode de Ward ou pour la sortie des $K$-moyennes
}
