% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{horn_parallele}
\alias{horn_parallele}
\title{Critère parallèle de Horn}
\usage{
horn_parallele(n, p, niveau = 0.95, nsim = 1000L)
}
\arguments{
\item{n}{nombre d'observations}

\item{p}{nombre de variables explicatives}

\item{niveau}{niveau pour le quantile}

\item{nsim}{nombre de simulations}
}
\value{
un vecteur de quantiles pour les valeurs propres
}
\description{
Cette fonction calcule par Monte Carlo les quantiles de la
distribution des valeurs propres pour une matrice de corrélation
de données indépendantes de dimension p, par le biais de tirages
d'une loi de Wishart
}
