% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{junkmail}
\alias{junkmail}
\title{Pourriels}
\format{
Une base de données avec 4601 lignes et 59 variables:
\describe{
  \item{\code{test}}{variable binaire pour données test et entraînement}
  \item{\code{make}}{mot}
  \item{\code{address}}{mot}
  \item{\code{all}}{mot}
  \item{\code{_3d}}{chaîne de caractères}
  \item{\code{our}}{mot}
  \item{\code{over}}{mot}
  \item{\code{remove}}{mot}
  \item{\code{internet}}{mot}
  \item{\code{order}}{mot}
  \item{\code{mail}}{mot}
  \item{\code{receive}}{mot}
  \item{\code{will}}{mot}
  \item{\code{people}}{mot}
  \item{\code{report}}{mot}
  \item{\code{addresses}}{mot}
  \item{\code{free}}{mot}
  \item{\code{business}}{mot}
  \item{\code{email}}{mot}
  \item{\code{you}}{mot}
  \item{\code{credit}}{mot}
  \item{\code{your}}{mot}
  \item{\code{font}}{mot}
  \item{\code{_000}}{chaîne de caractères}
  \item{\code{money}}{mot}
  \item{\code{hp}}{mot}
  \item{\code{hpl}}{mot}
  \item{\code{george}}{mot}
  \item{\code{_650}}{chaîne de caractères}
  \item{\code{lab}}{mot}
  \item{\code{labs}}{mot}
  \item{\code{telnet}}{mot}
  \item{\code{_857}}{chaîne de caractères}
  \item{\code{data}}{mot}
  \item{\code{_415}}{chaîne de caractères}
  \item{\code{_85}}{chaîne de caractères}
  \item{\code{technology}}{mot}
  \item{\code{_1999}}{mot}
  \item{\code{parts}}{mot}
  \item{\code{pm}}{mot}
  \item{\code{direct}}{mot}
  \item{\code{cs}}{mot}
  \item{\code{meeting}}{mot}
  \item{\code{original}}{mot}
  \item{\code{project}}{mot}
  \item{\code{re}}{mot}
  \item{\code{edu}}{mot}
  \item{\code{table}}{mot}
  \item{\code{conference}}{mot}
  \item{\code{semicolon}}{point virgule}
  \item{\code{paren}}{parenthèses}
  \item{\code{bracket}}{crochets}
  \item{\code{exclamation}}{point d'exclamation}
  \item{\code{dollar}}{signe de dollar}
  \item{\code{pound}}{signe de livre sterling}
  \item{\code{capavg}}{longueur moyenne des mots en majuscule}
  \item{\code{caplong}}{longueur des mots en majuscule}
  \item{\code{captotal}}{nombre total de majuscules}
  \item{\code{class}}{indicateur binaire, \code{1} pour un pourriel, \code{0} sinon}
}
}
\source{
George Forman et Hewlett-Packard, aide SAS
}
\usage{
junkmail
}
\description{
Données tirées de l'aide SAS servant à la classification de pourriels.
}
\keyword{datasets}
