% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistique.R
\name{lift}
\alias{lift}
\title{Courbe lift}
\usage{
lift(prob, resp)
}
\arguments{
\item{prob}{vecteur de probabilités estimées}

\item{resp}{variable réponse binaire}
}
\description{
Pour un vecteur de probabilitées (prédictions) d'un modèle, \code{prob},
et une variable binaire \code{resp}, calculer la courbe lift.
}
