% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{lnh}
\alias{lnh}
\title{Résultats de la LNH, 2021-2022}
\format{
Une base de données avec 1312 lignes et 6 variables:
\describe{
  \item{\code{date}}{[Date] date de la partie}
  \item{\code{domicile}}{[facteur] nom de l'équipe jouant à domicile}
  \item{\code{visiteur}}{[facteur] nom de l'équipe en visite}
  \item{\code{but_domicile}}{[entier] nombre de buts de l'équipe à domicile}
  \item{\code{but_visiteur}}{[entier]  nombre de buts de l'équipe en visite}
  \item{\code{type}}{[facteur] type de partie, soit \code{regulier}, soit temps supplémentaire si le match va en prolongation (\code{temps supp.}, soit \code{tirs au but} si l'égalité est brisée en tirs de barrage).}
}
}
\source{
Données extraites de \url{https://www.hockey-reference.com/leagues/NHL_2022_games.html} le 21 octobre 2022.
}
\usage{
lnh
}
\description{
Cette base de données contient les résultats des 32 équipes de la Ligue national de hockey (LNH) pour chaque partie avec le nombre de but par équipe.
}
\keyword{datasets}
