% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{lnh_BT}
\alias{lnh_BT}
\title{Résultats de la LNH, format Bradley-Terry}
\format{
Une base de données avec 1312 lignes et 32 variables:
\describe{
  \item{\code{vainqueur}}{[logique] variable binaire, \code{TRUE} si l'équipe à domicile remporte la partie}
  \item{\code{Arizona Coyotes}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Boston Bruins}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Buffalo Sabres}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Calgary Flames}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Carolina Hurricanes}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Chicago Blackhawks}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Colorado Avalanche}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Columbus Blue Jackets}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Dallas Stars}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Detroit Red Wings}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Edmonton Oilers}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Florida Panthers}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Los Angeles Kings}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Minnesota Wild}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Montreal Canadiens}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Nashville Predators}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{New Jersey Devils}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{New York Islanders}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{New York Rangers}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Ottawa Senators}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Philadelphia Flyers}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Pittsburgh Penguins}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{San Jose Sharks}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Seattle Kraken}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{St. Louis Blues}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Tampa Bay Lightning}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Toronto Maple Leafs}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Vancouver Canucks}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Vegas Golden Knights}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Washington Capitals}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
  \item{\code{Winnipeg Jets}}{[entier] 1 si l'équipe joue à domicile, -1 si elle joue à l'extérieur, 0 sinon.}
}
}
\usage{
lnh_BT
}
\description{
Transformation des données \link{lnh} pour l'ajustement d'un modèle de Bradley-Terry. La base de données ne contient que l'indicateur de l'équipe et une variable binaire indiquant l'équipe gagnante.
}
\author{
Léo Belzile
}
\keyword{datasets}
