% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{logistclient}
\alias{logistclient}
\title{Promotion de produits}
\format{
Une base de données avec 1000 lignes et 4 variables:
\describe{
 \item{\code{promo}}{variable binaire, 1 si le client s’est prévalue d’une offre promotionnelle, 0 sinon}
 \item{\code{nachats}}{nombre d’achats au magasin dans le dernier mois}
 \item{\code{sexe}}{variable binaire, 0 pour les femmes, 1 pour les hommes}
 \item{\code{tclient}}{variable catégorielle, soit \code{frequent} pour les clients réguliers ou \code{occasionnel} autrement}
}
}
\usage{
logistclient
}
\description{
Données simulées pour un cas fictif de promotion de produits.
}
\keyword{datasets}
