% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{logit6}
\alias{logit6}
\title{Données catégorielles et ordinales}
\format{
Une base de données avec 100 lignes et 3 variables:
\describe{
 \item{\code{x}}{âge du répondant (en années)}
 \item{\code{y1}}{quelle a été votre destination vacances l’année dernière: Québec (\code{0}), États-Unis (\code{1}) ou ailleurs (\code{2})?}
 \item{\code{y2}}{combien de fois êtes-vous allé au cinéma l’année dernière: moins de 5 fois (\code{1})), entre 5 et 10 fois (\code{2}), ou plus de 10 fois (\code{3}).}
}
}
\usage{
logit6
}
\description{
Cette base de données fictives contient 100 observations obtenues par voie de sondage auprès d’adultes âgés de 18 à 45 ans.
}
\keyword{datasets}
