% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{manquantes}
\alias{manquantes}
\alias{missing1}
\title{Recommendation du PRCA (données manquantes)}
\format{
Une base de données avec 500 lignes et 7 variables:
\describe{
 \item{\code{x1}}{type d'emploi occupé, un parmi à la maison (\code{1}), employé (\code{2}), ventes/services (\code{3}), professionnel (\code{4}) ou agriculture/ferme (\code{5})}
 \item{\code{x2}}{revenu familial annuel, un parmi \code{1}: moins de 25 000, \code{2}: 25 000 à 39 999, \code{3}: 40 000 à 59 999, \code{4}: 60 000 à 79 999 et \code{5}: 80 000 et plus}
 \item{\code{x3}}{sexe, soit homme (\code{0}) ou femme (\code{1})}
 \item{\code{x4}}{avez-vous déjà fréquenté une université? soit oui (\code{1}) ou non (\code{0})}
 \item{\code{x5}}{âge (en années)}
 \item{\code{x6}}{combien de fois avez-vous assisté à un rodéo au cours de la dernière année? soit \code{1}: 10 fois ou plus, \code{2}: entre six et neuf fois et \code{3}: cinq fois ou moins}
 \item{\code{y}}{seriez-vous intéressé à acheter un produit recommandé par le PRCA, soit oui (\code{1}) ou non (\code{0})}
}
}
\usage{
manquantes
}
\description{
Les auteurs cherchent à voir si le fait qu’un produit soit recommandé par le Professional Rodeo Cowboys Association (PRCA) a un effet sur les intentions d’achats.
}
\references{
Daneshvary, R. et Schwer, R. K. (2000) \emph{The Association Endorsement and Consumers’ Intention to Purchase}. Journal of Consumer Marketing \bold{17}, 203-213.
}
\author{
Denis Larocque
}
\keyword{datasets}
