% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{moyenne_modeles_stepAIC}
\alias{moyenne_modeles_stepAIC}
\title{Moyenne de modèles avec sélection séquentielle}
\usage{
moyenne_modeles_stepAIC(data, formula, aic = TRUE, B = 100L, ks = 2)
}
\arguments{
\item{data}{base de données de type \code{data.frame}}

\item{formula}{formule du modèle maximal considéré}

\item{aic}{logique; si \code{TRUE}, retourner le modèle qui retourne le plus petit critère d'Akaike}

\item{B}{entier, le nombre de réplications}

\item{ks}{valeur de la pénalisation pour le dernier modèle de \code{stepAIC}}
}
\value{
une liste avec
\itemize{
\item{\code{coefs}}{vecteur des coefficients}
\item{\code{nselect}}{nombre de coefficients non-nuls}
}
}
\description{
Cette fonction crée une copie du jeu de données par
autoamorçage nonparamétrique. La procédure \code{stepAIC}
de \pkg{MASS} est utilisée avec le modèle maximal spécifié dans
\code{formula} est considéré. Le modèle sélectionné est celui avec la plus
petite valeur du critère d'information selon la pénalité \code{ks};
la valeur par défaut correspond au critère d'Akaike.
}
