% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{multinom}
\alias{multinom}
\title{Satisfaction par rapport à un produit}
\format{
Une base de données avec 1000 lignes et 5 variables:
\describe{
 \item{\code{y}}{échelle de Likert de 1 à 5 représentant le score de satisfaction}
 \item{\code{sexe}}{sexe de l’individu, homme (\code{0}) ou femme (\code{1})}
 \item{\code{educ}}{niveau d’éducation le plus élevé complété ; secondaire (\code{sec}), collégial (\code{cegep}) ou universitaire (\code{uni})}
 \item{\code{revenu}}{variable catégorielle indiquant le revenu, soit faible (\code{1}), moyen (\code{2}) ou élevé (\code{3})}
 \item{\code{age}}{âge de l’individu (en années)}
}
}
\usage{
multinom
}
\description{
La satisfaction est mesurée à l’aide d’une échelle de Likert, allant de très insatisfait (1) à très satisfait (5).
}
\keyword{datasets}
