% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistique.R
\name{perfo_logistique}
\alias{perfo_logistique}
\title{Performance d'un modèle logistique}
\usage{
perfo_logistique(prob, resp)
}
\arguments{
\item{prob}{vecteur de probabilités estimées}

\item{resp}{variable réponse binaire}
}
\value{
une base de données avec les composantes suivantes:
\itemize{
\item{\code{coupe}}{points de coupe}
\item{\code{VP}: }{vrais positifs}
\item{\code{VN}: }{vrais négatifs}
\item{\code{FP}: }{faux positifs}
\item{\code{FN}: }{faux négatifs}
\item{\code{pcorrect}: }{taux de bonne classification}
\item{\code{sensi}: }{sensibilité ou rappel}
\item{\code{speci}: }{spécificité}
\item{\code{fpos}: }{taux de faux positifs}
\item{\code{fneg}: }{taux de faux négatifs}
\item{\code{prec}: }{précision}
\item{\code{F1}: }{score F1, moyenne géométrique du rappel et de la précision}
}
}
\description{
Calculer les statistiques d'ajustement (sensibilité, spécificité, taux de bonne classification, etc.) en fonction de points de coupures
}
