% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{predict.regsubsets}
\alias{predict.regsubsets}
\title{Prédiction pour regsubsets}
\usage{
\method{predict}{regsubsets}(object, newdata, id, ...)
}
\arguments{
\item{object}{objet de classe \code{regsubsets}}

\item{newdata}{base de données (\code{data.frame}) contenant les variables explicatives}

\item{id}{[entier] numéro du modèle ajusté par \code{regsubsets} pour lequel obtenir des prédictions}
}
\value{
un vecteur de prédictions
}
\description{
Cette fonction calcule les prédictions d'un modèle linéaire ajusté
contenant \code{id} variables et multiplie les coefficients par la matrice du modèle
pour retourner des prédictions.
}
\author{
R. Jordan Crouser
}
