% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistique.R
\name{predvc}
\alias{predvc}
\title{Prédictions par validation croisée répétée}
\usage{
predvc(modele, data = NULL, K = 10L, nrep = 10L, type = NULL)
}
\arguments{
\item{modele}{un modèle ajusté de type \code{lm} ou \code{glm}}

\item{data}{\code{NULL} une base de données \code{data.frame} si la liste \code{modele} n'inclut pas \code{data} parmi ses éléments}

\item{K}{entier, nombre de plis pour la validation croisée}

\item{nrep}{nombre de réplications}

\item{type}{type de prédiction; si \code{NULL}, la valeur par défaut est employée. S'assurer que la valeur est à une échelle logique pour le calcul de moyennes.}
}
\value{
vecteur de prédictions
}
\description{
Le modèle est ajusté de manière répété sur chaque pli à l'aide de \code{update} avec \code{predict}.
Le code est conçu pour les modèles linéaires généralisés, mais devrait fonctionner avec tout modèle qui définit des génériques S3.
}
\author{
Léo Belzile
}
