% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroupements.R
\name{rand}
\alias{rand}
\title{Indice de Rand}
\usage{
rand(x, y)
}
\arguments{
\item{x}{vecteur d'étiquettes A}

\item{y}{vecteur d'étiquettes B}
}
\description{
Calcul de la similarité; chaque paire est réduite à un index binaire (0 si les observations sont des regroupements différents, 1 si elles sont classées dans le même). Ces paires sont comptabilisées dans un tableau de contingence; l'indice de Rand est le taux de bonne classification.
}
