% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroupements.R
\name{rc_hclust}
\alias{rc_hclust}
\title{Critères de sélection}
\usage{
rc_hclust(x, hc, kmax = 10L)
}
\arguments{
\item{x}{base de données}

\item{hc}{objet de classe \code{hclust}}

\item{kmax}{nombre maximal de regroupements}
}
\value{
une base de données avec éléments
\itemize{
\item{\code{index}}{nombre de regroupements}
\item{\code{Rc}}{critère du R-carré}
\item{\code{Rcsp}}{critère du R-carré semi-partiel}
}
}
\description{
Fonction pour calculer le R carré et le
 R carré semi-partiel pour la procédure de
 regroupements hiérarchique (méthode de Ward)
}
