% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{regroupements1}
\alias{regroupements1}
\title{Analyse de regroupement}
\format{
Une base de données avec 150 lignes et 6 variables:
\describe{
 \item{\code{x1}}{échelle, activités sociales}
 \item{\code{x2}}{échelle, politiques de l’opérateur et références}
 \item{\code{x3}}{échelle, flexibilité des horaires}
 \item{\code{x4}}{échelle, santé et sécurité}
 \item{\code{x5}}{échelle, matériel publicitaire}
 \item{\code{x6}}{échelle, réputation}
}
}
\usage{
regroupements1
}
\description{
Données simulées inspirées de Hsu et Lee (2002) pour illustrer l'analyse de regroupement. Les données proviennent de trois
groupes, qui sont facilement discernables lorsque représentés
graphiquement à l'aide de composantes principales.

Les buts de l’étude étaient
\itemize{
\item Regrouper les gens de 55 et plus qui participent à des voyages organisés en autobus en groupes homogènes selon des caractéristiques reliées au choix de l’opérateur et du voyage.
\item Examiner les caractéristiques de ces groupes.
\item Examiner les caractéristiques démographiques de ces groupes.
}
}
\references{
Hsu, C. H. C. et Lee E.-J. (2002). \emph{Segmentation of Senior Motorcoach Travelers}. Journal of Travel Research, \bold{40}, 364-373.
}
\author{
Denis Larocque
}
\keyword{datasets}
