% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{renfe}
\alias{renfe}
\title{Prix de billets de train}
\format{
Une base de données de 10000 observations contenant les variables suivantes:
\describe{
  \item{prix}{prix du billet (en euros)}
  \item{dest}{facteur, indicateur binaire du trajet, soit de Barcelone vers Madrid ou de Madrid vers Barcelone}
  \item{tarif}{facteur, variable catégorielle indiquant le tarif du billet, un parmi \code{AdultoIda}, \code{Promo} et \code{Flexible}}
  \item{classe}{facteur classe du billet, soit  \code{Preferente}, \code{Turista}, \code{TuristaPlus} ou \code{TuristaSolo}}
  \item{type}{facteur variable catégorielle indiquant le type de train, \code{AVE-TGV}, \code{AVE} ou \code{REXPRESS}.}
  \item{duree}{longueur annoncée du trajet (en minutes)}
  \item{jour}{facteur entier indiquant le jour de la semaine du départ allant de dimanche (\code{1}) à samedi (\code{7})}
}
}
\source{
Kaggle \url{https://www.kaggle.com/thegurusteam/spanish-high-speed-rail-system-ticket-pricing}
}
\usage{
renfe
}
\description{
Un sous-ensemble de billets de trains pour le trajet entre Madrid et Barcelone
vendus par la Renfe pour trois types de trains, soit Alta Velocidad Española (\code{AVE}),
soit Alta Velocidad Española conjointement avec TGV (un partenariat entre la SNCF et Renfe
pour les trains à destination ou en provenance de Toulouse) \code{AVE-TGV}, soit les trains régionaux \code{REXPRESS};
 seuls les trains étiquetés \code{AVE} ou \code{AVE-TGV} sont des trains à grande vitesse.
}
\keyword{datasets}
