% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{salaireprof}
\alias{salaireprof}
\title{Discrimination salariale dans un collège}
\format{
Une base de données de 397 observations contenant les variables suivantes
\describe{
\item{\code{annees}}{nombre d'années depuis l'obtention du doctorat.}
\item{\code{service}}{nombre d'années de service.}
\item{\code{salaire}}{salaire de professeurs pendant l'année académique 2008--2009 (en milliers de dollars USD).}
\item{\code{echelon}}{échelon académique, soit adjoint (\code{adjoint}), aggrégé (\code{aggrege}) ou titulaire (\code{titulaire}).}
\item{\code{domaine}}{variable catégorielle indiquant le champ d'expertise du professeur, soit appliqué (\code{applique}) ou théorique (\code{theoretique}).}
\item{\code{sexe}}{indicateur binaire pour le sexe, \code{homme} ou \code{femme}.}
}
}
\source{
Fox J. and Weisberg, S. (2019) \emph{An R Companion to Applied Regression}, Third Edition, Sage.
}
\usage{
salaireprof
}
\description{
Une étude a été effectuée dans un collège américain pour explorer
la distribution des salaires et vérifier s'il existe des
inégalités salariales entre hommes et femmes.
}
\seealso{
\link[carData]{Salaries}
}
\keyword{datasets}
