% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistique.R
\name{select_pcoupe}
\alias{select_pcoupe}
\title{Sélection du point de coupure selon gain}
\usage{
select_pcoupe(
  modele,
  c00 = 1,
  c11 = 1,
  c01 = 0,
  c10 = 0,
  plot = FALSE,
  nrep = 10L,
  ncv = 10L
)
}
\arguments{
\item{modele}{modèle linéaire généralisé pour variables binaires}

\item{c00}{coût de classification pour vrai négatif}

\item{c11}{coût de classification pour vrai positif}

\item{c01}{coût de classification pour faux positif}

\item{c10}{coût de classification pour faux négatif}

\item{plot}{booléen, si \code{TRUE}, produit un graphique du lift. La valeur par défaut est \code{FALSE}.}

\item{nrep}{entier, nombre de réplications pour la validation croisée}

\item{ncv}{entier, nombre de groupes pour la validation croisée}
}
\value{
un graphique de la performance moyenne en fonction du point de coupure et une liste avec les points de coupure \code{ptcoupe}, la valeur optimale du point de coupure \code{opt} et la performance \code{performance}
}
\description{
La fonction calcule le gain pour une variable réponse binaire et calcule le gain (par défaut, le taux de bonne classification). Elle retourne le point de coupure optimal.
La fonction prend comme argument un modèle de classe
\code{glm} et réajuste le modèle de manière répétée pour
calculer la performance par validation croisée.
Le modèle est réajusté \code{nrep} fois.
}
