% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{selection1_train}
\alias{selection1_train}
\title{Sélection du degré d'un polynôme}
\format{
Une base de données avec 100 lignes et 2 variables:
\describe{
 \item{\code{x}}{variable explicative}
 \item{\code{y}}{variable réponse}
}
}
\usage{
selection1_train
}
\description{
Un jeu de données simulées issues d'un modèle polynomial d'ordre inconnu de la forme\deqn{y = \sum_{j=0}^{k} \beta_{j}x^{j}.} Le vrai modèle est cubique.
}
\author{
Denis Larocque
}
\keyword{datasets}
