% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{sondage}
\alias{sondage}
\alias{survey_sans_missing}
\title{Données de sondage sans valeurs manquantes}
\format{
Une base de données avec 1842 lignes et 10 variables:
\describe{
 \item{\code{wrkstat}}{facteur, occupation}
 \item{\code{marital}}{facteur, statut marital}
 \item{\code{childs}}{nombre d'enfants}
 \item{\code{age}}{âge (en années)}
 \item{\code{sex}}{facteur, sexe de l'individu}
 \item{\code{rincome}}{facteur, tranche de revenus}
 \item{\code{service}}{variable binaire, réponse à la question "avez-vous utilisé nos services dans le dernier mois?", soit oui (\code{1}) ou non (\code{0})}
 \item{\code{satisf}}{Si vous avez utilisé nos services, indiquez votre niveau de satisfaction, sur une échelle de 1 à 5 où 1 est très insatisfait et 5 très satisfait.}
}
}
\source{
Julie Meloche
}
\usage{
sondage
}
\description{
Base de données de service à la clientèle avec valeurs manquantes
}
\keyword{datasets}
