% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{sondage_entreprise}
\alias{sondage_entreprise}
\title{Sondage de satisfaction en entreprise}
\format{
Une base de données avec 482 lignes et 10 variables:
\describe{
 \item{\code{q8}}{La compagnie se préoccupe de la société}
 \item{\code{q9}}{La compagnie est dotée d’un grand sens éthique}
 \item{\code{q10}}{La compagnie est un meneur dans l’industrie des TI}
 \item{\code{q11}}{La compagnie est innovante}
 \item{\code{q12}}{La compagnie a des dirigeants forts, expérimentés et compétent}
 \item{\code{q13}}{La compagnie a une base financière saine}
 \item{\code{q14}}{Je peux faire confiance à la compagnie}
 \item{\code{q15}}{La compagnie a des publicités que j’apprécie particulièrement}
 \item{\code{q16}}{La compagnie est reconnue pour le bon traitement de ses employés}
 \item{\code{q17}}{La compagnie fait sa part pour aider la société}
}
}
\usage{
sondage_entreprise
}
\description{
Une compagnie fait un sondage auprès de ses employés. Les questions Q8–Q17 retenues pour
l’analyse factorielle dans la base de données sont mesurées à l’aide d’échelles de Likert allant de 1
(fortement en désaccord) à 5 (fortement d’accord).
}
\details{
Source inconnue, données fournies par Julie Meloche
}
\keyword{datasets}
