% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{survie4}
\alias{survie4}
\title{Temps d'abonnement (risques compétitifs)}
\format{
Une base de données avec 500 lignes et 6 variables:
\describe{
 \item{\code{temps}}{temps (en semaines) d'abonnement au service de téléphonie cellulaire}
 \item{\code{censure}}{entier, \code{1} si le temps est censuré (l’individu est toujours abonné à notre service), \code{2}/\code{3} si l'individu a quitté pour aller chez le compétiteur A/B, \code{4} si l'individu a quitté parce qu'il n'a plus besoin de cellulaire}
 \item{\code{age}}{âge du client au début de l’abonnement}
 \item{\code{sexe}}{variable binaire, soit femme (\code{1}) ou homme (\code{0})}
 \item{\code{region}}{région où habite le client en ce moment (valeurs entre 1 et 5)}
 \item{\code{service}}{nombre de services en plus du cellulaire auquel le client est abonné parmi internet, téléphone fixe, télévision (câble ou antenne parabolique).}
}
}
\usage{
survie4
}
\description{
Une entreprise oeuvrant dans le secteur des télécommunications s’intéresse aux facteurs influençant le temps qu’un client reste abonné à son service de téléphone cellulaire.
}
\author{
Denis Larocque
}
\keyword{datasets}
