% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{survival2}
\alias{survival2}
\title{Profils de clients}
\format{
Une base de données avec 4 lignes et 2 variables:
\describe{
 \item{\code{sexe}}{variable binaire, soit femme (\code{1}) ou homme (\code{0})}
 \item{\code{age}}{âge du client au début de l’abonnement}
}
}
\usage{
survival2
}
\description{
Quatre profils de clients pour le calcul de courbes de survie.
}
\author{
Denis Larocque
}
\keyword{datasets}
