% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{tisupport}
\alias{tisupport}
\title{Billets pour soutien TI}
\format{
Une base de données avec 279 observations et 5 variables:
\describe{
  \item{\code{duree}}{[entier] nombre d'heures (jours ouvrables) entre l'ouverture du billet et la fermeture ou la période de fin de collecte}
  \item{\code{statut}}{[facteur] statut de la requête, soit \code{ouvert} ou \code{ferme}}
  \item{\code{type}}{[facteur] type de demande, soit requête pour services (\code{requete}), soit un problème informatique ou bogue (\code{probleme})}
  \item{\code{niveau}}{[entier] niveau de difficulté assigné, une valeur plus élevée requiert des compétences plus importantes}
  \item{\code{ncommentaires}}{[entier] nombre de commentaires dans le suivi}
}
}
\usage{
tisupport
}
\description{
Cette base de données contient des données simulées pour des temps de résolutions de problèmes TI, selon le niveau de complexité de la requête, le type de demande.
}
\keyword{datasets}
