% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{visaechelles}
\alias{visaechelles}
\title{Analyse factorielle des données bancaires}
\format{
Une base de données avec 1027 lignes et 8 variables:
\describe{
 \item{\code{carvp}}{possession de la carte \code{Visa} Premier, soit oui (\code{1}) ou non (\code{0})}
 \item{\code{sexe}}{sexe, \code{0} pour homme, \code{1} pour femme}
 \item{\code{ech1}}{niveau d'activité}
 \item{\code{ech2}}{niveau d'endetemment}
 \item{\code{ech3}}{fortune}
 \item{\code{ech4}}{degré d'utilisation du compte}
 \item{\code{ech5}}{ancienneté}
 \item{\code{id}}{identifiant du client}
}
}
\usage{
visaechelles
}
\description{
Analyse en composantes principales des données \link{visaprem}; cinq échelles ont été créées. Cette base de données sert à la segmentation des clients.
}
\keyword{datasets}
