% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{visalogist}
\alias{visalogist}
\title{Données bancaires (régression logistique)}
\format{
Une base de données avec 1027 lignes et 24 variables:
\describe{
 \item{\code{matric}}{identifiant client}
 \item{\code{carvp}}{possession de la carte \code{Visa} Premier, soit oui (\code{1}) ou non (\code{0})}
 \item{\code{endet}}{taux d'endettement}
 \item{\code{facanl}}{log du montant facturé dans l'année en francs}
 \item{\code{itavcl}}{log du total des avoirs sur tous les comptes}
 \item{\code{kvunb}}{nombre de comptes à vue}
 \item{\code{moyrvl}}{log de la moyenne des mouvements nets créditeurs des 3 mois en milliers de francs}
 \item{\code{nbjdl}}{log du nombre de jours à débit des trois derniers mois}
 \item{\code{opgnbl}}{log du nombre d'opérations par guichet dans le mois}
 \item{\code{qcredl}}{log de la moyenne des mouvements créditeurs en milliers de francs}
 \item{\code{qsmoy}}{moyenne des soldes moyens sur 3 mois}
 \item{\code{relat}}{ancienneté de la relation (en mois)}
 \item{\code{sexe}}{sexe, \code{0} pour homme, \code{1} pour femme}
 \item{\code{uemnb}}{nombre de produits épargne monétaire}
 \item{\code{viemtl}}{log du montant des produits contrats vie en francs}
 \item{\code{vienb}}{nombre de produits contrats vie}
 \item{\code{xlgmtl}}{log du montant des produits d'épargne logement en francs}
 \item{\code{xlgnb}}{nombre de produits d'épargne logement}
 \item{\code{ylvmtl}}{log du montant des comptes sur livret en francs}
 \item{\code{ylvnb}}{nombre de comptes sur livret}
 \item{\code{zocnbr}}{nombre d'opérations par cartes}
 \item{\code{gagel}}{log du total des engagements en francs}
 \item{\code{role}}{facteur, étiquette pour données d'entraînement, de validation et de test}
}
}
\usage{
visalogist
}
\description{
Base de données de relations bancaires (voir \link{visacredm}) transformées pour le chapitre sur la régression logistique
}
\seealso{
\link{visacredm}
}
\keyword{datasets}
