% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorielle.R
\name{alphaC}
\alias{alphaC}
\title{Alpha de Cronbach}
\usage{
alphaC(x, na.rm = FALSE)
}
\arguments{
\item{x}{matrice de variables numériques}

\item{na.rm}{logique; si \code{TRUE}, les valeurs manquantes sont enlevées avant le calcul}
}
\value{
alpha de Cronbach
}
\description{
Calcul de la mesure de cohérence interne
pour la création d'échelles
}
