% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.R
\docType{data}
\name{electionsqc}
\alias{electionsqc}
\title{Élections générales québécoises}
\format{

}
\source{
Données ouvertes extraites du site du Directeur général des élections du Québec, licence \url{https://www.dgeq.org/licence.html}.
}
\usage{
electionsqc
}
\description{
Résultat électoraux pour les élections de 2018 et 2022 par circonscription en format large (parti, année). Les informations sociodémographiques sont basées sur les données du Recensement 2021 de Statistique Canada.
}
\details{
Comprend des données ouvertes octroyées sous la licence d'utilisation des données ouvertes du directeur général des élections disponible à l'adresse Web \url{dgeq.org}. L'octroi de la licence n'implique aucune approbation par le directeur général des élections de l'utilisation des données ouvertes qui en est faite.
}
\keyword{datasets}
